requirements.txt
fastapi
uvicorn
openai
python-dotenv
pytesseract
pillow
pymupdf
pandas
openpyxl
